refactoring websocket events into their own files
