+++ server/events/handleClose.js
@@ -1,0 +1,7 @@
+module.exports = {
+    handleClose(ws) {
+        ws.on('close', (code, reason) => {
+            console.log('closed', code, reason.toString())
+        })
+    }
+}
