+++ server/events/handleError.js
@@ -1,0 +1,7 @@
+module.exports = {
+    handleError(ws) {
+        ws.on('error', (e) => {
+            console.log('error', e)
+        })
+    }
+}
