+++ server/events/handleOpen.js
@@ -1,0 +1,7 @@
+module.exports = {
+    handleOpen(ws) {
+        ws.on('open', () => {
+            console.log('Connection to websocket open')
+        })
+    }
+}
