+++ server/events/index.js
@@ -1,0 +1,4 @@
+exports.handleError = require('./handleError').handleError
+exports.handleOpen = require('./handleOpen').handleOpen
+exports.handleClose = require('./handleClose').handleClose
+exports.handleMessage = require('./handleMessage').handleMessage
